Refactors out the use of GCS so a different object store can be used
Google Cloud Storage (GCS) isn't the only object store available and
unlike S3 there isn't a handy drop-in replacement like Minio available
to replace it. By refactoring out the use of GCS into a standard go
interface the entire server becomes much more modular for other
organizations to use.
